Move AutocompleteField under components/forms
