+++ server/webapp/WEB-INF/rails/webpack/views/components/forms/autocomplete.scss
@@ -16,4 +16,3 @@
 
 @import "../../global/common.scss";
 @import "./_awesomplete-import.scss";
-
