+++ server/webapp/WEB-INF/rails/webpack/views/pages/pipelines/components.scss
@@ -16,7 +16,7 @@
 
 @import "../../global/common";
 @import "task_terminal";
-@import "../../components/autocomplete/index";
+@import "../../components/forms/autocomplete";
 
 $diagram-width: 485px;
